Split artifact generation into a separate example
